Renamed IconImporter to IconProvider
